# cmd/laas/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.APICollectionResponse:
    properties:
      data:
        $ref: '#/definitions/models.APICollection'
      status:
        example: 200
        type: integer
    type: object
...
  models.AuditResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Audit'
        type: array
      paginationmeta:
        $ref: '#/definitions/models.PaginationMeta'
      status:
        example: 200
        type: integer
    type: object
...
  models.ChangeLogResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.ChangeLog'
        type: array
      paginationmeta:
        $ref: '#/definitions/models.PaginationMeta'
      status:
        example: 200
        type: integer
    type: object
...
  models.LicenseError:
    properties:
      error:
        example: invalid request body
        type: string
      message:
        example: invalid request body
        type: string
      path:
        example: /api/v1/licenses
        type: string
      status:
        example: 400
        type: integer
      timestamp:
        example: "2023-12-01T10:00:51+05:30"
        type: string
    type: object
...
info:
  contact:
    email: fossology@fossology.org
    name: FOSSology
    url: https://fossology.org
  description: Service to host license information for other services to query over
    REST API.
  license:
    name: GPL-2.0-only
    url: https://github.com/fossology/LicenseDb/blob/main/LICENSE
  title: laas (License as a Service) API
  version: 0.0.9
paths:
  /apiCollection:
    get:
      consumes:
      - application/json
      description: Returns the apis which require authentication and which do not
      operationId: getAPICollection
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APICollectionResponse'
        "500":
          description: Unable to parse swagger docs
          schema:
            $ref: '#/definitions/models.LicenseError'
      summary: Returns the apis which require authentication and which do not
      tags:
      - API Collection
  /audits:
    get:
      consumes:
      - application/json
      description: Get all audit records from the server
      operationId: GetAllAudit
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of records per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Audit records
          schema:
            $ref: '#/definitions/models.AuditResponse'
        "404":
          description: Not changelogs in DB
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get audit records
      tags:
      - Audits
  /audits/{audit_id}:
    get:
      consumes:
      - application/json
      description: Get a specific audit records by ID
      operationId: GetAudit
      parameters:
      - description: Audit ID
        in: path
        name: audit_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuditResponse'
        "400":
          description: Invalid audit ID
          schema:
            $ref: '#/definitions/models.LicenseError'
        "404":
          description: No audit entry with given ID
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get an audit record
      tags:
      - Audits
  /audits/{audit_id}/changes:
    get:
      consumes:
      - application/json
      description: Get changelogs of an audit record
      operationId: GetChangeLogs
      parameters:
      - description: Audit ID
        in: path
        name: audit_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChangeLogResponse'
        "400":
          description: Invalid audit ID
          schema:
            $ref: '#/definitions/models.LicenseError'
        "404":
          description: No audit entry with given ID
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: unable to find changes
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get changelogs
      tags:
      - Audits
  /audits/{audit_id}/changes/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific changelog of an audit record by its ID
      operationId: GetChangeLogbyId
      parameters:
      - description: Audit ID
        in: path
        name: audit_id
        required: true
        type: string
      - description: Changelog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChangeLogResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.LicenseError'
        "404":
          description: No changelog with given ID found
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get a changelog
      tags:
      - Audits
...
  /licenses:
    get:
      consumes:
      - application/json
      description: Filter licenses based on different parameters
      operationId: FilterLicense
      parameters:
      - description: SPDX ID of the license
        in: query
        name: spdxid
        type: string
      - description: License detector type
        in: query
        name: detector_type
        type: integer
      - description: GPLv2 compatibility flag status of license
        in: query
        name: gplv2compatible
        type: boolean
      - description: GPLv3 compatibility flag status of license
        in: query
        name: gplv3compatible
        type: boolean
      - description: Mary done flag status of license
        in: query
        name: marydone
        type: boolean
      - description: Active license only
        in: query
        name: active
        type: boolean
      - description: OSI Approved flag status of license
        in: query
        name: osiapproved
        type: boolean
      - description: FSF Free flag status of license
        in: query
        name: fsffree
        type: boolean
      - description: Copyleft flag status of license
        in: query
        name: copyleft
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit of responses per page
        in: query
        name: limit
        type: integer
      - description: External reference parameters
        in: query
        name: externalRef
        type: string
      - default: shortname
        description: Sort by field
        enum:
        - spdx_id
        - shortname
        - fullname
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Asc or desc ordering
        enum:
        - asc
        - desc
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Filtered licenses
          schema:
            $ref: '#/definitions/models.LicenseResponse'
        "400":
          description: Invalid value
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Filter licenses
      tags:
      - Licenses
    post:
      consumes:
      - application/json
      description: Create a new license in the service
      operationId: CreateLicense
      parameters:
      - description: New license to be created
        in: body
        name: license
        required: true
        schema:
          $ref: '#/definitions/models.LicenseDB'
      produces:
      - application/json
      responses:
        "201":
          description: New license created successfully
          schema:
            $ref: '#/definitions/models.LicenseResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/models.LicenseError'
        "409":
          description: License with same shortname already exists
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: Failed to create license
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Create a new license
      tags:
      - Licenses
  /licenses/{shortname}:
    get:
      consumes:
      - application/json
      description: Get a single license by its shortname
      operationId: GetLicense
      parameters:
      - description: Shortname of the license
        in: path
        name: shortname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LicenseResponse'
        "404":
          description: License with shortname not found
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get a license by shortname
      tags:
      - Licenses
    patch:
      consumes:
      - application/json
      description: Update a license in the service
      operationId: UpdateLicense
      parameters:
      - description: Shortname of the license to be updated
        in: path
        name: shortname
        required: true
        type: string
      - description: Update license body (requires only the fields to be updated)
        in: body
        name: license
        required: true
        schema:
          $ref: '#/definitions/models.LicenseUpdateJSONSchema'
      produces:
      - application/json
      responses:
        "200":
          description: License updated successfully
          schema:
            $ref: '#/definitions/models.LicenseResponse'
        "400":
          description: Invalid license body
          schema:
            $ref: '#/definitions/models.LicenseError'
        "404":
          description: License with shortname not found
          schema:
            $ref: '#/definitions/models.LicenseError'
        "409":
          description: License with same shortname already exists
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: Failed to update license
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Update a license
      tags:
      - Licenses
  /licenses/export:
    get:
      description: Export all licenses as a json file
      operationId: ExportLicenses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LicenseDB'
            type: array
        "500":
          description: Failed to fetch Licenses
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Export all licenses as a json file
      tags:
      - Licenses
  /licenses/import:
    post:
      consumes:
      - multipart/form-data
      description: Import licenses by uploading a json file
      operationId: ImportLicenses
      parameters:
      - description: licenses json file list
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ImportLicensesResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.LicenseImportStatus'
                  type: array
              type: object
        "400":
          description: input file must be present
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Import licenses by uploading a json file
      tags:
      - Licenses
  /licenses/preview:
    get:
      consumes:
      - application/json
      description: Get shortnames of all active licenses from the service
      operationId: GetAllLicensePreviews
      parameters:
      - description: Active license only
        in: query
        name: active
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LicensePreviewResponse'
        "400":
          description: Invalid active value
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: Unable to fetch licenses
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get shortnames of all active licenses
      tags:
      - Licenses
  /login:
    post:
      consumes:
      - application/json
      description: Login to get JWT token
      operationId: Login
      parameters:
      - description: Login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            properties:
              token:
                type: string
            type: object
      summary: Login
      tags:
      - Users
  /obligation_maps/license/{license}:
    get:
      consumes:
      - application/json
      description: Get obligation maps for a given license shortname
      operationId: GetObligationMapByLicense
      parameters:
      - description: Shortname of the license
        in: path
        name: license
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ObligationMapResponse'
        "404":
          description: No license with given shortname found or no map for
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get maps for a license
      tags:
      - Obligations
  /obligation_maps/topic/{topic}:
    get:
      consumes:
      - application/json
      description: Get obligation maps for a given obligation topic
      operationId: GetObligationMapByTopic
      parameters:
      - description: Topic of the obligation
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ObligationMapResponse'
        "404":
          description: No obligation with given topic found or no map for
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get maps for an obligation
      tags:
      - Obligations
  /obligation_maps/topic/{topic}/license:
    patch:
      consumes:
      - application/json
      description: Add or remove licenses from obligation map for a given obligation
        topic
      operationId: PatchObligationMap
      parameters:
      - description: Topic of the obligation
        in: path
        name: topic
        required: true
        type: string
      - description: Shortnames of the licenses with action
        in: body
        name: shortname
        required: true
        schema:
          $ref: '#/definitions/models.LicenseMapShortnamesInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ObligationMapResponse'
        "400":
          description: Invalid json body
          schema:
            $ref: '#/definitions/models.LicenseError'
        "404":
          description: No license or obligation found.
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: Failure to insert new maps
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Add or remove licenses from obligation map
      tags:
      - Obligations
    put:
      consumes:
      - application/json
      description: Replaces the license list of an obligation topic with the given
        list in the obligation map.
      operationId: UpdateLicenseInObligationMap
      parameters:
      - description: Topic of the obligation
        in: path
        name: topic
        required: true
        type: string
      - description: Shortnames of the licenses to be in map
        in: body
        name: shortnames
        required: true
        schema:
          $ref: '#/definitions/models.LicenseShortnamesInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ObligationMapResponse'
        "400":
          description: Invalid json body
          schema:
            $ref: '#/definitions/models.LicenseError'
        "404":
          description: No license or obligation found.
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Change license list
      tags:
      - Obligations
  /obligations:
    get:
      consumes:
      - application/json
      description: Get all active obligations from the service
      operationId: GetAllObligation
      parameters:
      - description: Active obligation only
        in: query
        name: active
        required: true
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of records per page
        in: query
        name: limit
        type: integer
      - default: asc
        description: Asc or desc ordering
        enum:
        - asc
        - desc
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerObligationResponse'
        "404":
          description: No obligations in DB
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get all active obligations
      tags:
      - Obligations
    post:
      consumes:
      - application/json
      description: Create an obligation and associate it with licenses
      operationId: CreateObligation
      parameters:
      - description: Obligation to create
        in: body
        name: obligation
        required: true
        schema:
          $ref: '#/definitions/models.ObligationDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SwaggerObligationResponse'
        "400":
          description: Bad request body
          schema:
            $ref: '#/definitions/models.LicenseError'
        "409":
          description: Obligation with same body exists
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: Unable to create obligation
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Create an obligation
      tags:
      - Obligations
  /obligations/{topic}:
    delete:
      consumes:
      - application/json
      description: Deactivate an obligation
      operationId: DeleteObligation
      parameters:
      - description: Topic of the obligation to be updated
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: No obligation with given topic found
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Deactivate obligation
      tags:
      - Obligations
    get:
      consumes:
      - application/json
      description: Get an active based on given topic
      operationId: GetObligation
      parameters:
      - description: Topic of the obligation
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerObligationResponse'
        "404":
          description: No obligation with given topic found
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Get an obligation
      tags:
      - Obligations
    patch:
      consumes:
      - application/json
      description: Update an existing obligation record
      operationId: UpdateObligation
      parameters:
      - description: Topic of the obligation to be updated
        in: path
        name: topic
        required: true
        type: string
      - description: Obligation to be updated
        in: body
        name: obligation
        required: true
        schema:
          $ref: '#/definitions/models.ObligationUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerObligationResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.LicenseError'
        "404":
          description: No obligation with given topic found
          schema:
            $ref: '#/definitions/models.LicenseError'
        "500":
          description: Unable to update obligation
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - ApiKeyAuth: []
      summary: Update obligation
      tags:
      - Obligations
...
  /obligations/export:
    get:
      description: Export all obligations as a json file
      operationId: ExportObligations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ObligationDTO'
            type: array
        "500":
          description: Failed to fetch obligations
          schema:
            $ref: '#/definitions/models.LicenseError'
      security:
      - '{}': []
        ApiKeyAuth: []
      summary: Export all obligations as a json file
      tags:
      - Obligations
...
  /oidc/login:
    get:
      consumes:
      - application/json
      description: Handles the OIDC Login
      operationId: OidcLogin
      produces:
      - application/json
      responses:
        "307":
          description: Temporary redirect to the oidc provider login page
          schema:
            type: string
        "500":
          description: Failed to login
          schema:
            $ref: '#/definitions/models.LicenseError'
      summary: Handles the OIDC Login
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    description: Token from /login endpoint
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"